Restructure autopilot folder and autopilot class
