--- motion/autopilot/src/PID/__init__.py
@@ -1,1 +1,0 @@
-from .PIDregulator import PIDRegulator
