--- motion/autopilot/src/backstepping/__init__.py
@@ -1,1 +1,0 @@
-from .backstepping_control import BacksteppingDesign, BacksteppingControl
