+++ motion/motion_launch/launch/motion.launch
@@ -1,13 +1,18 @@
 <?xml version="1.0"?>
 <launch>
 
+  <!-- Guidance nodes-->
   <node pkg="los_guidance" type="los_guidance_backstepping.py" name="los_guidance" if="$(eval type == 'real')"/>
   <node pkg="los_guidance" type="los_guidance_backstepping.py" name="los_guidance" output="screen" if="$(eval type == 'simulator')"/>
 
-  <node pkg="thruster_interface"  type="thruster_interface_node.py"   name="thruster_interface"/>
-  <node pkg="vortex_allocator"    type="vortex_allocator_node"        name="allocator"/>
+  <!-- Controller nodes-->
+  <node pkg="autopilot"           type="autopilot.py"                 name="autopilot"            output="screen"/>
   <node pkg="dp_controller"       type="dp_controller_node"           name="controller"           output="screen"/>
   <node pkg="interface"           type="controller_interface.py"      name="controller_interface" output="screen"/>
 
+  <!-- Thruster nodes-->
+  <node pkg="thruster_interface"  type="thruster_interface_node.py"   name="thruster_interface"/>
+  <node pkg="vortex_allocator"    type="vortex_allocator_node"        name="allocator"/>
   <node pkg="mcu_interface"       type="mcu_interface_node"           name="mcu_interface"  if="$(eval type == 'real')"/> 
+
 </launch>
